[feat]: 추가 구현을 위한 라우팅 설정, 가독성 향상
